+++ semana15/Projeto/JSON/teachers.json
@@ -1,0 +1,1 @@
+[]
