+++ semana15/Projeto/src/Classes/DaytimeClass.ts
@@ -1,0 +1,5 @@
+import { Class } from "./Class";
+
+export class DaytimeClass extends Class {
+  
+}
