+++ semana15/Projeto/src/Classes/Student.ts
@@ -1,0 +1,12 @@
+import { User } from "./UserInterface";
+import moment = require("moment");
+
+export class Student implements User {
+  constructor(
+    public name: string,
+    public email: string,
+    public id: string,
+    private birthDate: moment.Moment,
+    public hobbies: string[]
+  ) { }
+}
